requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
apscheduler==3.10.4
httpx==0.27.2
pydantic-settings==2.5.2
slowapi==0.1.9
python-dotenv==1.0.1
